pyproject.toml
[project]
name = "jobcreation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "pymupdf>=1.27.2",
    "python-dotenv>=1.0.0",
    "pdfplumber>=0.11.9",
    "psycopg2>=2.9.11",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "click>=8.3.1",
    "colorama>=0.4.6",
    "lxml>=6.0.2",
    "networkx>=3.6.1",
    "numpy>=2.4.2",
    "packaging>=26.0",
    "pandas>=3.0.1",
    "pypdf>=6.7.4",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "six>=1.17.0",
    "tzdata>=2025.3",
    "urllib3>=2.6.3",
    "sqlalchemy>=2.0.48",
    "pydantic-settings>=2.13.1",
    "asyncpg>=0.31.0",
]
